# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quuta API v1",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "ログイン",
                "description": "メールアドレスとパスワードでログイン",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "認証情報をいれる",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "quuta@quuta.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "quuta12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "認証情報が間違っている",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/follower": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "フォロワー取得",
                "description": "ユーザのフォロワーを一覧取得する",
                "operationId": "getFollower",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_get_follower"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "クエリパラメタに誤りがある"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/following": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "フォロー中取得",
                "description": "ユーザがフォロー中のユーザを一覧取得する",
                "operationId": "getFollowing",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_get_following"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "クエリパラメタに誤りがある"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/follow": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "フォロー",
                "description": "ユーザをフォローする",
                "operationId": "store_follow",
                "requestBody": {
                    "$ref": "#/components/requestBodies/user_store_follow"
                },
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "リクエストボディに誤りがある"
                    },
                    "201": {
                        "description": "フォロー情報が登録された",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "フォロー解除",
                "description": "フォローを解除する",
                "operationId": "destroy_follow",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/user_destroy_follow"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "クエリパラメタに誤りがある"
                    },
                    "201": {
                        "description": "フォロー情報が削除された",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/post": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "投稿",
                "description": "新しい投稿をする",
                "operationId": "storeNewPost",
                "requestBody": {
                    "$ref": "#/components/requestBodies/post_store_request_body"
                },
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "リクエストボディに誤りがある"
                    },
                    "201": {
                        "description": "投稿が作成された",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/image": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "投稿画像アップロード",
                "description": "投稿時に使う画像をアップロードする。ここで返却される画像idを使って投稿する。",
                "operationId": "storeNewImage",
                "requestBody": {
                    "$ref": "#/components/requestBodies/post_store_image_request_body"
                },
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "422": {
                        "description": "リクエストボディに誤りがある"
                    },
                    "201": {
                        "description": "投稿が作成された",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/post/{id}": {
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "投稿削除",
                "description": "投稿を削除する",
                "operationId": "destoryPost",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/post_destory_post_id"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "403": {
                        "description": "ログインユーザに削除できない投稿が指定された"
                    },
                    "422": {
                        "description": "リクエストボディに誤りがある"
                    },
                    "204": {
                        "description": "投稿が削除された",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/home_timeline": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "ホームタイムライン取得",
                "description": "ログインユーザ用のタイムラインを取得する",
                "operationId": "getHomeTimeline",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/timeline_get_since_id"
                    },
                    {
                        "$ref": "#/components/parameters/timeline_get_until_id"
                    },
                    {
                        "$ref": "#/components/parameters/timeline_get_count"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user_timeline": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "ユーザタイムライン取得",
                "description": "特定のユーザごとの投稿を取得する",
                "operationId": "getUserTimeline",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/timeline_get_user_id"
                    },
                    {
                        "$ref": "#/components/parameters/timeline_get_since_id"
                    },
                    {
                        "$ref": "#/components/parameters/timeline_get_until_id"
                    },
                    {
                        "$ref": "#/components/parameters/timeline_get_count"
                    }
                ],
                "responses": {
                    "401": {
                        "description": "認証されていない"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/user/{id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "ユーザ情報取得",
                "description": "ユーザの登録情報を取得する",
                "operationId": "getUserInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "取得したいユーザのID"
                    }
                ],
                "responses": {
                    "404": {
                        "description": "指定されたユーザが存在しない"
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
            "post_destory_post_id": {
                "name": "id",
                "in": "path",
                "description": "削除する投稿のid",
                "required": true
            },
            "timeline_get_since_id": {
                "name": "since_id",
                "in": "query",
                "description": "投稿のid、これ以降に投稿された投稿を取得する",
                "example": "1"
            },
            "timeline_get_until_id": {
                "name": "until_id",
                "in": "query",
                "description": "投稿のid、これ以前に投稿された投稿を取得する",
                "example": "30"
            },
            "timeline_get_count": {
                "name": "count",
                "in": "query",
                "description": "取得する投稿の数、 デフォルト10、最大50",
                "example": "20"
            },
            "timeline_get_user_id": {
                "name": "user_id",
                "in": "query",
                "description": "取得するユーザのid",
                "required": true
            },
            "user_get_info": {
                "name": "id",
                "in": "path",
                "description": "取得するユーザのid"
            },
            "user_get_follower": {
                "name": "user_id",
                "in": "query",
                "description": "対象ユーザのid、このユーザのフォロワーを取得する。",
                "required": true
            },
            "user_get_following": {
                "name": "user_id",
                "in": "query",
                "description": "対象ユーザのid、このユーザがフォローしているユーザ取得する。",
                "required": true
            },
            "user_destroy_follow": {
                "name": "user_id",
                "in": "query",
                "description": "削除するユーザのid",
                "required": true
            }
        },
        "requestBodies": {
            "post_store_request_body": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "content",
                                "restaurant_id",
                                "image_ids"
                            ],
                            "properties": {
                                "content": {
                                    "description": "投稿の本文",
                                    "type": "string"
                                },
                                "restaurant_id": {
                                    "description": "投稿が紐づく店舗を表すid",
                                    "type": "string"
                                },
                                "image_ids": {
                                    "description": "投稿される写真のid。idは事前に画像をアップロードして取得する。",
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    }
                                },
                                "like_flag": {
                                    "description": "店舗にいいねをつけるかのフラグ",
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post_store_image_request_body": {
                "description": "投稿のid、これ以降に投稿された投稿を取得する",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "image"
                            ],
                            "properties": {
                                "image": {
                                    "description": "アップロードする画像データ",
                                    "type": "string",
                                    "format": "binary"
                                },
                                "dish_name": {
                                    "description": "画像の料理名",
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "user_store_follow": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "user_id"
                            ],
                            "properties": {
                                "user_id": {
                                    "description": "フォローするユーザのid",
                                    "type": "string"
                                },
                                "subscription_flag": {
                                    "description": "ユーザのアクションを購読するか",
                                    "type": "boolean"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearerで認証する"
            }
        }
    }
}